debian/console-conf.*.service: run after snapd.recovery-chooser-trigger.service
On Core20 devices, the user may decice to invoke a recovery chooser by holding
down a specific key. Make sure that the recovery chooser trigger detection
service runs before, so that by the time console-conf runs, the trigger
detection window is closed and we may launch the chooser if needed.

Note, the patch only includes the bits for ensuring the correct order during
boot.

Signed-off-by: Maciek Borzecki <maciek.borzecki@gmail.com>